requirements.txt
# requirements.txt (مستقر لِـ Apple Silicon)
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1

tensorflow-macos==2.15.0
tensorflow-metal==1.1.0

matplotlib==3.9.2
seaborn==0.13.2
ipykernel==6.29.5
jupyterlab==4.2.5

protobuf==4.25.3
h5py==3.11.0

opencv-python==4.10.0.84
mediapipe==0.10.14
Pillow==10.4.0
albumentations==1.4.8
imbalanced-learn==0.12.3
tqdm==4.66.4
kaggle==1.6.14
split-folders==0.5.1